+++ NEWS.md
@@ -1,7 +1,11 @@
 # NEWS
 
+
+
 ## merTools 0.2.2
 
+- Remove `plyr` and replace with `dplyr`
+- Fix issue #62 `varList` will now throw an error if `==` is used instead of `=`
 - Fix issue #54 `predictInterval` did not included random effects in calculations
   when `newdata` had more than 1000 rows and/or user specified `parallel=TRUE`. 
   Note: fix was to disable the `.paropts` option for `predictInterval` ... user